AdaptBarren.jl Documentation

AdaptBarren.jl Documentation

Functions

Index

Functions

AdaptBarren.Operator — Type

Operator(paulis::Array{Pauli{UInt64},1}, coeffs::Array{ComplexF64,1})

Construct an operator which is a weighted sum of Pauli strings.

AdaptBarren.Operator — Method

Operator(p::Pauli)

Create an operator that is a single Pauli string with weight 1.0.

AdaptBarren.Pauli — Type

Pauli{T<:Unsigned}

Pauli mask representation of a Pauli string. This type is parametric on T so that smaller sized integers can be used to represent the Pauli string, if possible. This also performs a check that the given integers constitute a valid Pauli string.

Examples

Pauli string: "IIXYZYZIYIZ"

id: 11000001010 -> 1546

x: 00100000000 -> 256

y: 00010100100 -> 164

z: 00001010001 -> 81

AdaptBarren.DeltaYStopper — Function

DeltaYStopper(delta::Float64 = 1e-8, n_best::Int64 = 5)

Halt the algorithm if the n_best many energies are all within delta of each other.

This is supposed to indicate that the algorithm has stalled in performance.

AdaptBarren.EnergyErrorPrinter — Method

EnergyErrorPrinter(gse::Float64)

Print the difference between the energy at the last iteration and the ground state energy, given as the argument gse.

AdaptBarren.EnergyPrinter — Method

EnergyPrinter()

Print the energy returned at the last iteration.

AdaptBarren.FloorStopper — Function

FloorStopper(floor::Float64, delta::Float64 = 1e-8)

Halt the algorithm if the current energy falls within delta of floor.

AdaptBarren.MaxGradientPrinter — Method

MaxGradientPrinter()

Print the maximum gradient at each layer.

AdaptBarren.MaxGradientStopper — Function

MaxGradientStopper(eps::Float64 = 1e-8)

Halts the algorithm if the maximum gradient falls below eps.

AdaptBarren.OperatorIndexPrinter — Function

OperatorIndexPrinter(formatted_ops = nothing)

At each layer, indicate the operator chosen by the ADAPT criterion.

If formatted_ops === nothing, the index of each operator as it exists in the pool will be printed. If a list of strings is given, the string of that index will instead be printed.

AdaptBarren.ParameterPrinter — Method

ParameterPrinter()

Prints the optimal parameters at each iteration.

AdaptBarren.ParameterStopper — Method

ParameterStopper(max_pars::Int64)

Halt the algorithm if the maximum number of parameters is exceeded.

AdaptBarren.clump_degenerate_values — Method

clump_degenerate_values(v::Vector{T}, degen_tol = 1e-10) where T<:Real

Given a sorted vector v, produce a vector of vectors where the resulting elements are all within degen_tol of each other in absolute value.

Examples

Given the vector [0, 1, 1, 2, 2, 2, 3], the clumped output should be [[0], [1, 1], [2, 2, 2], [3]].

AdaptBarren.commutator — Function

commutator(A::Operator, B::Operator, simplify::Bool = true)

Compute the commutator [A,B] for operators A and B.

AdaptBarren.diagonal_operator_to_vector — Method

diagonal_operator_to_vector(O::Operator)

Return the diagonal of an operator which is diagonal in the computational basis.

AdaptBarren.exp_val — Method

exp_val(A::Operator, state::Array{ComplexF64,1}, tmp::Array{ComplexF64})

Compute the expectation value <state|A|state>.

The tmp array is mutated, and state is left unchanged.

AdaptBarren.fast_grad! — Method

fast_grad!( ham::Operator, ansatz, pars, result, psi::Vector{ComplexF64}, sigma::Vector{ComplexF64}, tmp1::Vector{ComplexF64}, tmp2::Vector{ComplexF64}) where T <: Unsigned

Compute the gradient of the VQE cost function using an efficient method.

Counting matrix multiplications

The number of required matrix multiplications is 1 + N + |support(ham)| + N-1 + N*(1+1+1) = 2*N + 3*N = 5*N ~ O(N)

This could in principle be reduced to 3*N multiplications, but at this point I don't want to spend more time optimizing this.

Arguments

ham::Operator: Hamiltonian to compute the expectation value with respect to.

ansatz: List of generators for ansatz.

pars: Parameters used in the ansatz.

result: Vector to store the resulting gradient.

psi::Vector{ComplexF64}: Temporary vector.

sigma::Vector{ComplexF64}: Initial state for the ansatz.

tmp1::Vector{ComplexF64}: Temporary vector.

tmp2::Vector{ComplexF64: Temporary vector.

AdaptBarren.finite_difference! — Method

finite_difference!(ham::Operator, ansatz::Vector{Pauli{T}}, pars::Vector{Float64}, result::Vector{Float64}, initial_state::Vector{ComplexF64}, tmp1::Vector{ComplexF64}, tmp2::Vector{ComplexF64}, eps::Float64 = 1e-8) where T <: Unsigned

Compute the gradient of the VQE cost function using the finite difference method.

Arguments

ham::Operator: Hamiltonian to compute the expectation value with respect to.

ansatz: List of generators for ansatz.

pars: Parameters used in the ansatz.

result: Vector to store the resulting gradient.

initial_state::Vector{ComplexF64}: Initial state for the ansatz.

tmp1::Vector{ComplexF64}: Temporary vector.

tmp2::Vector{ComplexF64}: Temporary vector.

eps::Float64 = 1e-8: Small value to use for finite difference.

AdaptBarren.get_energy_gap — Method

get_energy_gap(h::Operator)

Compute the difference beteween the largest and smallest eigenvalues of a given Hamiltonian.

AdaptBarren.get_ground_state — Method

get_ground_state(h::Operator)

Compute the smallest eigenvalue of an operator.

Currently only diagonal operators are supported.

AdaptBarren.ground_state_overlap — Function

ground_state_overlap(hamiltonian::Vector{Float64}, state::Vector{ComplexF64}, degen_tol = 1e-10)

Compute the overlap between the given state and the ground state of a given operator.

Note that if the given operator has degenerate ground states, the overlaps are summed over. Numerically, all states with energy E such that |E - E_0| <= degen_tol are said to be degenerate.

AdaptBarren.matrix_to_operator — Method

matrix_to_operator(A::Array{ComplexF64,1})

Return the matrix A represented as an operator in the Pauli basis.

AdaptBarren.max_cut_hamiltonian — Method

max_cut_hamiltonian(n::Int, edges::Vector{Tuple{Int, Int, T}}) where T<:Real

Return the max cut Hamiltonian acting on n qubits which corresponds to the edge set edges.

Examples

Elements of the edge set look like (1,2,5.0) corresponding to an edge between vertices 1 and 2 with edge weight 5.0.

AdaptBarren.minimal_complete_pool — Method

minimal_complete_pool(n::Int64)

Return the minimal complete pool on n qubits corresponding with the V pool in the qubit-ADAPT paper.

AdaptBarren.num_qubits — Method

num_qubits(O::Operator)

Compute the number of qubits on which O acts non-trivially, including qubits with lower index than the maximum.

For example, a Pauli with Pauli string XYI would act on 3 qubits, but IYX would act on only 2 qubits.

AdaptBarren.num_qubits — Method

num_qubits(P::Pauli)

Determine the number of qubits on which P acts non-trivially.

AdaptBarren.one_local_pool_from_axes — Method

one_local_pool_from_axes(n::Int64, axes::Vector{Int64})

Return the pool on n qubits with single qubit operators corresponding to each element of axes.

Has size n * |axes|.

AdaptBarren.op_chop! — Function

op_chop!(A::Operator, tol::Float64 = 0.0)

Remove terms inplace from an operator whose coefficients have coefficients with magnitude less than tol.

AdaptBarren.op_product — Method

op_product(A::Operator, B::Operator)

Computes the product of operators A and B.

AdaptBarren.operator_to_matrix — Method

operator_to_matrix(O::Operator)

Compute the matrix representation of a given operator.

AdaptBarren.pauli_ansatz! — Method

pauli_ansatz!(axes, pars, result::Array{ComplexF64,1}, tmp::Array{ComplexF64,1})

Evolve result according to the given ansatz and angles.

Note that result is the initial state, evolved inplace, and is also the resulting state. tmp is also mutated, but should be considered junk.

Arguments

axes::Array{Pauli{T},1} where T<:UInt: The generators used in the ansatz.

pars::Array{Float64,1}: The angles of rotation for each of the generators.

result::Array{ComplexF64,1}: The state to be evolved.

tmp::Array{ComplexF64,1}: A temporary vector.

AdaptBarren.pauli_commute — Method

pauli_commute(P::Pauli, Q::Pauli)

Return whether Paulis P and Q compute.

AdaptBarren.pauli_mult! — Method

pauli_mult!(pm::Pauli, state::Array{ComplexF64,1}, result::Array{ComplexF64,1})

Compute the action of a Pauli string on a given statevector and store the result.

Note that the result is stored result and state remains unchanged.

AdaptBarren.pauli_product — Method

pauli_product(P::Pauli, Q::Pauli)

Compute the product of the operators P and Q and return the result.

AdaptBarren.pauli_string_to_pauli — Function

pauli_string_to_pauli(ps::String)

Create a Pauli object from a given Pauli string.

The rightmost element of the string is interpreted as the first-indexed qubit, i.e. IIXY has the Y operator acting on qubit 1.

AdaptBarren.pauli_to_pauli_string — Method

pauli_to_pauli_string(P::Pauli{T}, n::Int)

Create a Pauli string from the given Pauli.

AdaptBarren.qaoa_layer! — Method

qaoa_layer!(hamiltonian::Array{Float64,1}, gamma::Float64, result::Array{ComplexF64,1})

Given a diagonal operator (hamiltonian), evolve the state result by an angle gamma.

AdaptBarren.random_regular_max_cut_hamiltonian — Method

random_regular_max_cut_hamiltonian(n::Int, k::Int; rng = _DEFAULT_RNG, weighted = true)

Return a random Hamiltonian for a max cut problem on n qubits.

The corresponding graph is degree k. If an RNG is provided, this will be used to sample the graph and edge weights. If weighted is true, the edge weights will be randomly sampled from the uniform distribution U(0,1).

AdaptBarren.random_two_local_ansatz — Function

random_two_local_ansatz(n::Int64, k::Int64, axes=[0,1,2,3]; rng=_DEFAULT_RNG)

Produce a pool on n qubits consisting of k randomly sampled Pauli strings that are two local and determined by the elements of axes.

AdaptBarren.two_local_pool — Function

two_local_pool(n::Int64, axes=[0,1,2,3])

Returns the two local pool on n qubits using elements of axes as terms in the Pauli strings.

AdaptBarren.two_local_pool_from_pairs — Method

two_local_pool_from_pairs(n::Int64, pairs::Vector{Tuple{Int64, Int64}}; include_reverses = true)

Return the pool on n qubits with all two qubit Pauli strings defined in pairs.

Examples

The pairs vector [(1,1), (2,3)] corresponds to all X_i X_j and Y_i Z_j terms acting on all pairs of qubits. If include_reverses is true, then terms like Y_j Z_i are included.

AdaptBarren.uniform_state — Method

uniform_state(n::Int)

Return the vector representation of the all |+> state on n qubits.

Base.:* — Method

*(x::Operator, y::Operator)

Computes x*y for operators x and y.

Base.:+ — Method

+(x::Operator, y::Operator)

Computes x+y for operators x and y.

Base.:- — Method

-(x::Operator, y::Operator)

Computes x-y for operators x and y.

Index

AdaptBarren.Operator

AdaptBarren.Operator

AdaptBarren.Pauli

AdaptBarren.DeltaYStopper

AdaptBarren.EnergyErrorPrinter

AdaptBarren.EnergyPrinter

AdaptBarren.FloorStopper

AdaptBarren.MaxGradientPrinter

AdaptBarren.MaxGradientStopper

AdaptBarren.OperatorIndexPrinter

AdaptBarren.ParameterPrinter

AdaptBarren.ParameterStopper

AdaptBarren.clump_degenerate_values

AdaptBarren.commutator

AdaptBarren.diagonal_operator_to_vector

AdaptBarren.exp_val

AdaptBarren.fast_grad!

AdaptBarren.finite_difference!

AdaptBarren.get_energy_gap

AdaptBarren.get_ground_state

AdaptBarren.ground_state_overlap

AdaptBarren.matrix_to_operator

AdaptBarren.max_cut_hamiltonian

AdaptBarren.minimal_complete_pool

AdaptBarren.num_qubits

AdaptBarren.num_qubits

AdaptBarren.one_local_pool_from_axes

AdaptBarren.op_chop!

AdaptBarren.op_product

AdaptBarren.operator_to_matrix

AdaptBarren.pauli_ansatz!

AdaptBarren.pauli_commute

AdaptBarren.pauli_mult!

AdaptBarren.pauli_product

AdaptBarren.pauli_string_to_pauli

AdaptBarren.pauli_to_pauli_string

AdaptBarren.qaoa_layer!

AdaptBarren.random_regular_max_cut_hamiltonian

AdaptBarren.random_two_local_ansatz

AdaptBarren.two_local_pool

AdaptBarren.two_local_pool_from_pairs

AdaptBarren.uniform_state

Base.:*

Base.:+

Base.:-